requirements.txt
streamlit==1.49.0
Pillow==11.3.0
openai==1.60.0